관리자 페이지 기능 추가, 회원탈퇴, 유저 정보 수정 구현
- 관리자 페이지에서 삭제 시 본인 여부 확인
- 회원탈퇴, 비밀번호, 닉네임 변경, 회원가입 시 이메일 중복 체크, 비밀번호 조건 검사 구현
- 로그아웃 util화
- HTTP 메서드 별로 users DAO 패키지 분류

+++ src/main/java/com/visitJapan/controller/AdminController.java
@@ -13,9 +13,9 @@ import java.util.stream.Collectors;
 import org.bson.types.ObjectId;
 import org.json.JSONObject;
 
-import com.visitJapan.dao.users.AdminGrantDAO;
-import com.visitJapan.dao.users.DeleteUserDAO;
-import com.visitJapan.dao.users.GetAllUserDAO;
+import com.visitJapan.dao.users.delete.DeleteUserDAO;
+import com.visitJapan.dao.users.get.GetAllUserDAO;
+import com.visitJapan.dao.users.post.AdminGrantDAO;
 import com.visitJapan.dto.db.UsersDTO;
 
 
@@ -51,22 +51,33 @@ public class AdminController extends HttpServlet {
 	protected void doDelete(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		DeleteUserDAO deleteUserDAO = new DeleteUserDAO();
 		request.setCharacterEncoding("UTF-8");
+		response.setContentType("application/json; charset=UTF-8");
 	    
 	    String body = request.getReader().lines().collect(Collectors.joining()); // js 요청 본문
 	    JSONObject json = new JSONObject(body); // json 읽기
 	    
 	    String userId = json.getString("userId");
 	    ObjectId id = new ObjectId(userId);
+	    ObjectId myId = (ObjectId) request.getSession().getAttribute("id"); // 현재 로그인 중인 유저의 id
 	    
-	    boolean result = deleteUserDAO.removeUser(id);
-	    
-	    response.setContentType("application/json; charset=UTF-8");
-	    if (result) {
-	    		response.setStatus(HttpServletResponse.SC_NO_CONTENT); // 204
+	    if (id != myId) { // 자기 자신이 아니라면
+		    boolean result = deleteUserDAO.removeUser(id);
+		    
+		    if (result) {
+        			System.out.println(id + " 회원 탈퇴 완료");
+		    		response.setStatus(HttpServletResponse.SC_NO_CONTENT); // 204
+		    } else {
+		        response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR); // 500
+		        response.getWriter().write(
+		            "{\"status\":\"error\", \"message\":\"유저 삭제 실패\"}"
+		        );
+		    }
+		 
 	    } else {
-	        response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR); // 500
+	        response.setStatus(HttpServletResponse.SC_FORBIDDEN); // 403
+	        response.setContentType("application/json; charset=UTF-8");
 	        response.getWriter().write(
-	            "{\"status\":\"error\", \"message\":\"유저 삭제 실패\"}"
+	            "{\"status\":\"forbidden\", \"message\":\"본인은 삭제할 수 없습니다.\"}"
 	        );
 	    }
 	}
